Fixed memory leaks and refactored value interfaces.
Refactored how constants are managed by the Bytecode class so that they could be properly freed
(without being inadvertently garbage-collected). This involved some heavy refactoring in the
compiler.

Fixed missing deallocation code in the Allocator class. Now all managed values should be deleted
when the VM destructor is called.

Refactored value so that the Value union itself contains most of the needed functionality. Still
need to update the rest of the code base to use this.